requirements.txt
tqdm~=4.64.0
numpy~=1.22.3
pandas~=1.4.2
Pillow~=8.0.1
torch~=1.9.0
scikit-learn~=1.0.2